Don't replace the current history entry for a cross-origin initiator
In general, when a frame is navigated to its current url, we treat it
as a soft reload, and replace the current history entry rather than
appending a new one. However, when the navigation is initiated by a
cross-origin iframe, that behavior could potentially be used to
guess the current url by navigating the target to the url you think
it is at and observing how history.length changes. It's better to
forbid that information leak and say that a cross-origin-initiated
navigation to the current url always creates a new entry.

Bug: 1208614
Change-Id: Ib862c06952a3db61d2aa183c634edf6b5419c91a